prov/util: add FI_PEER support to util_cq
This patch adds FI_PEER support to the util_cq allowing any provider
that uses the util_cq to be used as either an owner or a peer.
All existing calls to directly write to the CQ are kept as is so any
provider wishing to leverage the FI_PEER support will have to be
modified to use the provided ofi_peer_cq_write and ofi_peer_cq_write_error
calls which go through the callbacks.

Summary of all the changes:
- Only allocate CQ resources if the provider is the owner, otherwise, set
the peer_cq using the context provided
- Add ofi_peer_cq_write and ofi_peer_cq_write_error to use peer callbacks
- Rename a few functions to stick with the convention that util_* calls are
called internally and ofi_* calls are exposed to the providers
- Squash ofi_cq_init and fi_cq_init which were separate and confusing

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>