prov/sm2: Create sm2 provider for development
Create an experimental sm2 provider to hold the development for the nemesis
shared memory algorithm. sm2 is based on the shm provider, but names
have been changed and all the protocols besides inject have been
removed.  This will allow us to quickly develop and test the new queue
mechanism, and we will restore functionality once we are confident it is
working.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>
Co-authored-by: Alexia Ingerson <alexia.ingerson@intel.com>
Co-authored-by: Seth Zegelstein <szegel@amazon.com>