prov/efa: make base_ep->src_addr an entity and initialize it properly
This patch change the type of "src_addr" in efa_base_ep from
a pointer to struct efa_ep_addr. This avoids allocating memory
for it, and casting it to proper type.

It also moved the initialization of src_addr to efa_base_ep_enable()
after qp is created, because this is the moment we have the
complete information to initialize address.

Signed-off-by: Wei Zhang <wzam@amazon.com>