prov/cxi: Locally registered MR do not require RKEY
For the previous client key implementation, the provider
required MR registered with only local access be assigned
a remotely accessible MR key. This was fixed so that MR
registered for local access only do not require a unique
key to be passed in the registration. In this case if
the MR is queried for the RKEY, FI_KEY_NOTAVAIL will be
returned.

NETCASSINI-4781

Signed-off-by: Steve Welch <welch@hpe.com>