prov/efa: Introduce RDMA Write Feature to EFA
Update documentation to describe RDMA-write based data transfer
to the EFA provider.  This is implemented as an extra feature
as described in the EFA protocol version 4.

Update the EFA handshake protocol to exchange the RDMA write
feature flag.

Add a function to detect if RDMA write is supported by the
domain endpoint.  Currently, it always returns false.

Signed-off-by: Luke Robison <lrbison@amazon.com>