prov/efa: Use shm provider's FI_AV_USER_ID support
The shm provider supports FI_AV_USER_ID flag. This flag
associates a user-assigned identifier with each av entry that is
returned with any completion entry in place of the AV's address.
EFA provider could use efa's fi_addr as the user id (passed as
as an input to the *fi_addr argument in fi_av_insert call)
With this change, there is no need to translate shm address back
to efa address when reading shm cq entry.

Signed-off-by: Shi Jin <sjina@amazon.com>