prov/efa: Set default value of use_device_rdma
use_device_rdma was not set to a default value
in rxr_env_get_use_device_rdma(). Hence when
the environment variable FI_EFA_USE_DEVICE_RDMA
is not set and the variable use_device_rdma is
not set, random behaviour occurs.

This patch fixes the bug by initializing
the default value of use_device_rdma

Signed-off-by: Vishwas Dsouza <vidsouza@amazon.com>