prov/tcp: Copy the remote sockaddr when connecting to peer
With io_uring, the sockaddr argument given to connect() must
live in memory until the completion event is polled. To do
so, the patch adds a memory allocation to store a copy
of the given sockaddr object when fi_connect() is executed.

This memory allocation is not needed for traditional sockets
but I wanted to keep the implementation identical for both
sockets and io_uring.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>