prov/shm: Store IPC availability in smr_domain
This patch serves as a prerequisite to an upcoming patch.
The logic of the function cuda_is_ipc_enabled() will be
changed to fetch the value of an environment variable.
With that change the calls to ofi_hmem_is_ipc_enabled()
in smr_generic_rma() and smr_generic_sendmsg() will be
costly.

This patch eliminates this problem by storing the values
IPC availbility in smr_domain structure.

Signed-off-by: Vishwas Dsouza <vidsouza@amazon.com>