prov/tcp: Replace fabric->progress with eq->progress
From issue #8505:
With the new TCP provider, polling an EQ results in progressing all
the domains that belong to the same fabric + the fabric itelsef.
This is what the function xnet_progress_all() does. Now, if we
 have multiple threads polling some EQs from the same fabric, that
would result in a high contention against the fabric's lock
fabric->util_fabric.lock and the domain's lock progress->active_lock.

Part of the problem here is that we progress all domains when reading
from an EQ, but not all domains will necessarily generate events on
that EQ.  We only need to progress domains that will generate events
on the EQ.  This can be determined by which EPs on a domain are bound
to that EQ.  In the case of DNN, the app assigns different EPs, CQs,
and EQ to a single thread, such that there's a 1:1 correlation
between a domain and an EQ.

As a first step, remove the fabric->progress engine and instead
replace it with an eq->progress engine.  This works better for
multi-threaded apps that have an EQ per thread, and actually makes
more sense, since we drive progress through the EQ.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>