configure: Fix the handling of some '--enable-<feature>' options
With the '--enable-<feature>' option, it is possible to disable the
feature by passing '--enable-<feature>=no' or '--disable-<feature>'.
The handling of some of such options (e.g. --enable-cuda-dlopen)
doesn't take that into account and may incorrectly enable the feature
when shouldn't.

Fix by checking the value when the option is given before enabling the
feature.

Also fix a few places where a prefix is added to only one side of string
comparison.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>