prov/tcp: Fix segfault accessing invalid fi_addr
In xnet_match_tag_addr(), we lookup the tag queue based on the peer->
fi_addr.  However, there's no guarantee that the peer has been
inserted into the AV.  This occurs during certain DAOS tests.  As
a result, the tag queue is invalid, and trying to walk it can
result in accessing invalid memory.

Add a check for a valid peer->fi_addr prior to checking for matches
on the src queue.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>