prov/hook/dmabuf_peer_mem: Handle IPC handle caching in L0
In older oneAPI L0 library IPC handles are not cached so that the
returned dmabuf fds need to be explicitly closed to avoid running
out of file descriptors.

In newer oneAPI L0 libraries IPC handles are cached so that it
becomes not only unnecessary but also errorous to explicitly close
the fds.

Detect the IPC handle caching status at runtime so that the dmabuf
fds can be correctly handled.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>