fabtests/rdm_stress: Remove count from json test configs
The count parameter does not work generically when paired with
message requests.  For each request, we need a corresponding
response for the test to make progress.  However, we can't assume
that the server will queue thousands of requests before we start
looking for a response.  Flow control can back up the sending
of requests.

Remove the count from the config files.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>