core/hmem: Allow specifying Ze devices as device_handles
Level zero has the concept of drivers and devices, with devices
belonging to driver objects.  As a result, two devices can have
the same index as long as they are from different drivers.
To disambiguate what device might be referred to, extend the
API to allow specifying the ze device.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>