src/hmem_ze: update ZE to support new driver index specification
API was updated to include a way to specify driver in addition to device index
by using a new call - fi_hmem_ze_device which will turn a driver and device index
into a single int to be passed into the registration call. This int will eventually
be passed into the ze copy and get handle calls which currently assume it is just
a device index. This patch updates the ZE code to make sure we only use the device
index, and'ing out the driver index. This also adds an assert because currently only
one driver is supported. More updates are necessary to support multiple drivers.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>