prov/tcp: Add missing call to xnet_submit_uring()
When canceling an outstanding sockctx, the function needs to call
xnet_progress_uring() in order to progress the cancel request.
The problem is that xnet_progress_uring() may prep new SQEs
that are unrelated to the canceled request but never submitted.

As a fix, make sure that xnet_submit_uring() is called before leaving
xnet_uring_cancel().

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>