prov/tcp: Mark the EP as disconnected before flushing the queues
With io_uring, outstanding SQEs are canceled in xnet_ep_flush_all_queues().
This operation is asynchronous and CQEs can be completed while the
function waits for the canceled SQE to complete. It causes an issue
where xnet_ep_disable() can be called in a nested way like in the
following callstack:

xnet_ep_disable()
  xnet_ep_flush_all_queues()
    xnet_uring_cancel()
      xnet_progress_uring()
        xnet_ep_disable()       <-- post EQ event

The problem is that the second call xnet_ep_disable() posts
the EQ event because the first call to xnet_ep_disable() marks
the EP state as XNET_DISCONNECTED too late (i.e., after the call
to xnet_ep_flush_all_queues()).

The fix implemented in the patch ensures that the EP state is set
to XNET_DISCONNECTED before entering xnet_ep_flush_all_queues().

A similar fix was implemented in xnet_ep_close() because RDM
applications are allowed to close the EP without shutting it
down.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>