man/fi_peer: fix and clarify context descriptions
user_context was split into two context - owner_context and
peer_context for the owner and peer to use, respectively.
This updates references to user_context to the applicable new
context and adds some more clarification to the usage of the
rx_entry fields, especially the contexts.

Also fixes numbering

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>