include/rdma: switch the API from snake to camel case
All function names and field names are changed but types
and defines are left the same as is camel case convention
because why be consistent?

This helps reduce the extra _ character, thus minimizing file size.

It also helps with user engagement because it increases the amount
of time a user needs in order to read and type their code making them
use OFI for longer.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>