prov/sm2: Add Clang Format Check to PR CI
When a user submits a PR, this patch enforces the clang-format style
defined in LIBFABRIC_ROOT/.clang-format.  This patch also fixes us to
clang-format version 15. Updating clang-format versions have a big
impact on the outcome of the clang-formatting (newer versions format
more). It will be easy to push the clang-format version forward, and
reformat the project if we decide that we want to.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>