prov/tcp: Fix a use-after if ep closed after reject
The following sequence causes a use-after-free that can be
detected by ASAN:
- Incoming connection request received on PEP
- Create the endpoint with fi_endpoint(info)
- Reject the connection request with fi_reject(pep, info->handle);
- Close the endpoint with fi_close(&ep->fid)

The use-after free occurs in tcpx_ep_close() when the following
condition is evaluated:
if (ep->fid && ep->fid->fclass == TCPX_CLASS_CM)

The pointer stored in ep->fid references a memory that is freed.
When tcpx_endpoint() is executed the EP saves a reference
to the handle in ep->handle. Later, this handle is freed
by fi_reject() and ep->handle becomes a dangling pointer as
a consequence.
Finally, tcpx_ep_close() accesses ep->fid, which triggers
the use-after-free (note that ep->fid and ep->handle are part
of the same union).

As a fix, tcpx_ep_close() now accesses ep->fid only when the ep
state is different than TCPX_RCVD_REQ.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>