prov/sm2: Add Deleivery Complete Support to SM2
This change has performance degredations.  We add 80 bits to the header
for every message transmission. We add another if check to the critical
path, and two more assignments.

Add context pointer + original op_flags to the header of the message.
When sender receives its FQE back from the receiver, if
FI_DELIEVERY_COMPLETE is in the op_flags, we write a delivery complete
completion.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>