prov/shm: store incoming message size in tagged rx_entry.
Currently, when getting an unexpected tagged msg, shm does not fill the
size of incoming message to the rx_entry before queueing it. However, middlewares
like MPI will call fi_trecvmsg with FI_PEEK flag in MPI_Probe() to peek if there
is an incoming message, and application will call MPI_Get_count to derive the size
of the incoming message from the output of MPI_Probe(). In this case, shm needs to
store the incoming message size in tagged rx_entry before queueing it, so owner provider
can get the size and write it in the CQE for FI_PEEK.

Signed-off-by: Shi Jin <sjina@amazon.com>