prov/efa: update rx_entry.total_len in srx queue_tag.
Currently, when peer provider getting an unexpected tagged msg, efa incorrectly set the total_len
in the unexp rx entry as 0 because get_tag API does not have size as the input.
However, middlewares like MPI will call fi_trecvmsg with FI_PEEK flag in MPI_Probe() to peek if there
is an incoming message, and application will call MPI_Get_count to derive the size
of the incoming message from the output of MPI_Probe(). In this case, efa needs to
store the incoming message size in tagged rx_entry before queueing it, so efa
can get the size and write it in the CQE after finding the rx_entry in the unexp queue.

The value of peer_rx_entry.size will be updated by peer provider before calling queue_tag.

Signed-off-by: Shi Jin <sjina@amazon.com>