fabtests/component: Add ze-dlopen functions to component tests
Component tests directly use level-zero functions which causes a linking issue.
If you want to build on a system with the headers but not the libraries you get
dependency issues with all fabtests.
This adds all of the functions required by component tests to the ze_ops and links
them as dl. It also exposes them in the header and changes how the component
tests access those functions.
Doing this allows libze_loader to no longer be a static dependency of fabtests.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>