core: Fix segfault in ZE GPU TX slow path
ZE GPU is_addr_valid() function does not correctly
handle the zeMemGetAllocProperties() function return
code. When the address is unknown it returns success
and set the memory properties to UNKNOWN. This would
lead to a segfault when the device was required (in
our slow path for TX and CXI_FORK_SAFE).

NETCASSINI-4047

Signed-off-by: Steve Welch <welch@hpe.com>