include/ofi_mem: smr_freestack_init enhancements
Correctly initialize fs->free in smr_freestack_init().
This is required for SM2 b/c it re-uses the SHM file, so we require
that we initialize to fs->free 0 every time.

Add new macro smr_freestack_isfull() to determine if freestack is full.
This is being added because SM2 can only delete itself from the shared
memory file header (for its slot to be re-used), if and only it has
every free queue entry returned to it. If its free queue entry isn't
full, and we allow another process to overwrite our region, we risk
corrupting  a peers queue that has a pointer to our FQE.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>