mr/cache: Add peer ID to MR cache
IPC uses MR cache to cache the IPC handles using the remote buffer virtual
address as part of the key. The find_within and find_overlap functions
only use the address to determine if the region matches.
However, when you have mutli-process communication as in all-to-all test,
it is entirely possible that the virtual addresses of all processes is the
same and that results in non-unique entries in the cache. The cache search
can then return the wrong entry for the peer. In the IPC case this would
return the wrong IPC handle, which would result in the cache entry being
invalidated and the new IPC handle opened again. This would cause a
significant performance hit.

To resolve this issue, add a peer ID entry in the mr key. In effect we create
a unique address space per peer, and in the process ensure we get the correct
entry returned.

Also updated the find_within and find_overlap to use the peer_ID.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>