core/hmem: introduce hmem_data field to memory registration
Introduce a wildcard field hmem_data to provide an extension point for
HMEM interfaces that require additional information to device id.

The immediate use case is to allow different providers to communicate
the presence of a GDR handle. This is achieved with a companion flag,
i.e. OFI_HMEM_DATA_GDRCOPY_HANDLE.

Furthermore, this change makes it possible to offload similar operations
to the user application, e.g. GDR pinning.

Signed-off-by: Wenduo Wang <wenduwan@amazon.com>