prov/linkx: rework memory registration
Reworked the memory registration structures. There are still a few
problems. We need to understand when exactly to do memory registration.
Currently OMPI MTL avoids doing any registration with CXI, but the BTL
does. There isn't a good way to figure out the difference between the two
paths. Some more thought needs to be put in how we do memory registration
at the LINKx level.

For one-sided:

The way memory registration works is P1 would register a piece of memory
and get back a key. P2 would do the same. P1 and P2 would exchange these
keys. When P1 wants to do an RMA from P2 memory, it passes P2 key to P2 in
the RMA operation. P2 would then use the key to find the memory and
complete the RMA operation.

This represents a problem with LINKx, because LINKx essentially abstracts
multiple core providers, and provides its own key back to the application.
However, when the LINKx Key is passed to the core provider, it makes no
sense to it.

What we need is a way for the core provider to ask LINKx to extract its
key given the LINKx key. To do that we will need to implement a separate
peer interface to enable this type of key lookup.

For now, hack LINKx to always expect exactly two providers. For one-sided
this would be sufficient for running on frontier, since we'll ever only link
SHM and CXI.

However, the above solution will need to be implemented for the generic
case.

Don't calloc() on the memory registration path. Instead create a buffer
pool and grab out of that when doing memory registration.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>