prov/efa: prepare local read pkt entry for in-order aligned send/recv.
On receiver side, we copy the data from host bounce buffer
to the application buffer via rdma-read. In-order aligned send/recv requires the
source data address to be 128 byte aligned. To that end,
we first copy the data from bounce buffer to pkt allocated
from read-copy pkt pool, which is 128 byte aligned for LL128.
Additionally, the read-copy pkt entry will store the data
at the beginning of pkt_entry->wiredata, which is also 128
byte aligned because the pkt_entry struct has size of 128 bytes.

Signed-off-by: Shi Jin <sjina@amazon.com>