prov/efa: Increase the pad size in rxr_pkt_entry
The current rxr_pkt_entry struct is 192 bytes for debug mode.
In-order aligned 128 byte ops requires the application data, which is stored in
pkt_entry->wiredata, to be 128 byte aligned. wiredata is the
last member in rxr_pkt_entry struct, which has an offset as
sizeof(struct rxr_pkt_entry) in the bufpool. To make this offset
128 aligned, we have to make sizeof(struct rxr_pkt_entry) to be
a multiple of 128 bytes. For non-debug mode, the size is already
128 bytes so no change is needed. For debug-mode, we need to
increase the pad size from 48 to 112 bytes to make the
sizeof(struct rxr_pkt_entry) to be 256 bytes.

Signed-off-by: Shi Jin <sjina@amazon.com>