prov/efa: remove RXR_RX_MAX_QUEUE_SIZE and RXR_TX_MAX_QUEUE_SIZE
RXR_TX_MAX_QUEUE_SIZE is not used any more, so this patch removed
it.

RXR_RX_MAX_QUEUE_SIZE is used to as max_cnt of ofi_bufpool_create().
Its value was set to UIN32_MAX, which is same as no limit.
Therefore this patch used 0 for max_cnt, which has the same effect
and removed RXR_RX_MAX_QUEUE_SIZE.

Signed-off-by: Wei Zhang <wzam@amazon.com>