mr/cache: Fix fi_mr_info initialization
In the ofi_mr_rbt_overlap() function an iovec pointer is passed in as
the key. However, it's then passed to util_mr_find_overlap() and
assigned to an fi_mr_info pointer. This assumes that only the iovec
pointer is accessed. With the addition of the peer_id check that is no
longer true.

To avoid this situation allocate an fi_mr_info block on the stack,
initialize the block to 0, copy in the iovec info, and pass it to the
call to ofi_rbmap_search()

Switch the initialization of fi_mr_info to the {0} form to make it
consistent throughout the code

Signed-off-by: Amir Shehata <shehataa@ornl.gov>