fabtests/rdm_tagged_peek: fix race condition synchronization
It was possible for the client to complete the finalize send before
the server gets the last 0xabc completion, causing a tag mismatch
when processing the incorrect receive.

This removes the extra synchronization sends and adds an OOB sync to
make sure all of the test sends have finished before quitting the test.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>