prov/cxi: Fix multi-recv byte counting
The previous counting down approach was prone to early freeing from the
parent request. The correct approach is to record the number of bytes
consumed when an auto unlink event occurs. Then, the parent will only be
freed once the sum of all child request data lengths equals the auto
unlink value.

NETCASSINI-5147

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>