prov/tcp: Handle case where epoll isn't natively supported
In the tcp provider, if wait support is requested, the code
calls ofi_dynpoll_get_fd() to retrieve the epoll fd.  That
fd can be passed into the progress epoll fd.  This works as
long as epoll support is available on the platform.

However, if epoll isn't available, ofi_dynpoll_get_fd() will
return an invalid fd.  We need to handle the non-epoll case
differently, as we can't rely on chaining the file descriptors
for event reporting.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>