core/epoll: Fix build warning for ofi_dynpoll_get_fd
If epoll is emulated, ofi_dynpoll_get_fd() is not usable and returns
a pointer, rather than an int, resulting in a build warning.

Update the epoll abstraction to add a helper for returning the fd.
In the case of emulated epoll, return invalid_socket.  This addresses
the build warning coming from calls to ofi_dynpoll_get_fd().

Signed-off-by: Sean Hefty <sean.hefty@intel.com>