prov/verbs: Remove saved_wc_list and use CQ directly
The provider reads completions from the HW CQ, processes them,
and if needed reports them up to the user in the appropriate
format.  Through these steps, it doesn't actually write entries
into the libfabric CQ.  However, there are times when it reads
from the HW CQ as part of internal progress, but can't report the
completions to the user yet.  Rather than writing an entry into
the libfabric CQ, it insteads stores the HW completion on a
separate list.

The entire process is confusing, results in extra memory,
leaves the CQ allocated memory unused, and increases code by
not properly leveraging the common CQ functions.

Remove the saved_wc_list and related memory pools.  Instead,
after reading from the HW CQ, simply write a completion into
the libfabric CQ if one is needed.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>