prov/verbs: Recover RXM connection from verbs QP in error state
Verbs QP can transition to error state for number of reasons
such as retransmission time out or protection error, rendering RXM
connection unusable until QP is back in RTS state.  Since not all
RDMA adapters can transition a QP from error state back to RTS,
the solution is to destroy the QP and create a new one on
subsequent use.

On a completion with error status, check current QP state (expensive)
against error state. Write a FI_SHUTDOWN event to event queue
for RXM to process and close the connection.  When a new request to
use the connection, a new QP will be created.

Signed-off-by: Chien Tin Tung <chien.tin.tung@intel.com>